+++ packages/domain-events/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.28.15](https://github.com/standardnotes/snjs/compare/@standardnotes/domain-events@2.28.14...@standardnotes/domain-events@2.28.15) (2022-05-30)
+
+**Note:** Version bump only for package @standardnotes/domain-events
+
+
+
+
+
 ## [2.28.14](https://github.com/standardnotes/snjs/compare/@standardnotes/domain-events@2.28.13...@standardnotes/domain-events@2.28.14) (2022-05-27)
 
 **Note:** Version bump only for package @standardnotes/domain-events